Add implementation that helps take in non-whole number MC mods *Example: LL4361Z

+++ src/main/java/seedu/duke/modules/Module.java
@@ -5,13 +5,13 @@ import seedu.duke.exceptions.ModuleException;
 public class Module {
     private String moduleCode;
     private String moduleGrade;
-    private int moduleMC;
+    private float moduleMC;
     private boolean moduleTaken;
     private int moduleDate;
 
     private String moduleDescription;
 
-    public Module(String moduleCode, int moduleMC, int moduleDate, String moduleDescription) {
+    public Module(String moduleCode, float moduleMC, int moduleDate, String moduleDescription) {
         this.moduleCode = moduleCode;
         this.moduleMC = moduleMC;
         this.moduleDate = moduleDate;
@@ -54,7 +54,7 @@ public class Module {
         this.moduleGrade = null;
     }
 
-    public int getModuleMC() {
+    public float getModuleMC() {
         return moduleMC;
     }
